+++ crates/cli/tests/reference/self-type.bg.js
@@ -15,7 +15,7 @@ export class Test {
     }
     constructor() {
         const ret = wasm.test_new();
-        this.__wbg_ptr = (ret >>> 0);
+        this.__wbg_ptr = ret >>> 0;
         TestFinalization.register(this, this.__wbg_ptr, this);
         return this;
     }
@@ -37,7 +37,7 @@ export class Test {
     }
 }
 if (Symbol.dispose) Test.prototype[Symbol.dispose] = Test.prototype.free;
-export function __wbg___wbindgen_throw_6ddd609b62940d55(arg0, arg1) {
+export function __wbg___wbindgen_throw_81fc77679af83bc6(arg0, arg1) {
     throw new Error(getStringFromWasm0(arg0, arg1));
 }
 export function __wbindgen_init_externref_table() {